refactor: move component to seperate file
