+++ src/components/App/Header/HamburgerButton.module.css
@@ -1,0 +1,39 @@
+.mHeaderButton {
+  @media (min-width: above-tablet) {
+    display: none;
+  }
+
+  position: absolute;
+  top: 0;
+  right: 0;
+  z-index: 9;
+  display: block;
+  transform: translate(0, 0);
+  padding: 9px 19px 0 19px;
+
+  &::before,
+  &::after,
+  span {
+    background: #fff;
+    border-radius: 3px;
+    content: '';
+    display: block;
+    width: 26px;
+    height: 2px;
+    margin: 7px 0;
+    transition: all 0.2s ease-out;
+    transform-origin: 50% 50%;
+  }
+
+  &.isNavOpen {
+    &::before {
+      transform: rotate(45deg) translate(5px, 8px);
+    }
+    &::after {
+      transform: rotate(-45deg) translate(5px, -8px);
+    }
+    span {
+      opacity: 0;
+    }
+  }
+}
